Add required version of ebookmaker as submodule
This repo reles on setanta's (Marcelo Lira's) ebookmaker, which is *not*
the version of ebookmaker available on pypi (that one is Project
Gutenberg's ebookmaker, which is *not* compatible).

Unfortunately, stanta's ebookmaker does not include license information
and seems to no longer be maintained. I've started looking at other
options for generating the epub file, but, in the meantime, I added
setanta's ebookmaker as a submodule to let this project work out of the
box.