Allow setting default options on RateLimited-wrapped funtions; support cafile as cli arg
NOT TESTED YET: Currently IP banned.
Need to test that pages work w/ ssl context added to all urlopen calls and that context and cafile don't mutually conflict (will require actually using the cafile option properly...currently, I've only tested that it gets passed to urlopen as written and can be overridden, not that it works when pointing to a real/valid CA file).